requirements.txt
alembic==0.8.9
boto==2.45.0
click==6.6
Flask==0.11.1
Flask-Cors==3.0.2
Flask-HTTPAuth==3.2.1
Flask-Migrate==2.0.2
Flask-OAuth==0.12
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.6
MarkupSafe==0.23
oauth2==1.9.0.post1
oauthlib==2.0.1
pkg-resources==0.0.0
PyMySQL==0.7.9
python-editor==1.0.3
requests==2.12.4
requests-oauthlib==0.7.0
six==1.10.0
SQLAlchemy==1.1.4
Werkzeug==0.11.11
